Split project-detail.tsx into source sub-components (410 → 210)
- project-detail.tsx              shell — fetch, instructions section,
                                  source list composition
- project-detail/types.ts         ProjectDetail + ProjectSource + the
                                  three-way AddSourceMode union
- project-detail/AddSourceCard    file / URL / text mode-switching add
                                  form, owns the multipart vs JSON
                                  branching internally
- project-detail/SourceCard       single source row with icon, label,
                                  optional URL/text/notes, remove button

The add-source form's local state (label, url, text, notes, mode,
submitting, file ref) all moves into AddSourceCard, leaving the shell
to own only project / instructions / addOpen.

All files under 220 lines.
